Add '*' parameter for language and literaltype column OPTIONS
This new option ignores the language tags and data types in literals
in FILTER expressions. This might slow down the SPARQL query a bit
but it is quite handy when the user is not sure what kind of language
tags or data types the literals have in a certain triple setting.