Bug fix for fdw_private data cache
This change makes sure that the fdw_private data is cachable by
storing them as Const instead of naked pointers. This fixes a bug
that was causing some queries to fail with a 'unrecognized node type'
error. The fdw_private data can now be copied by the planner and
therefore preserved.