Add prototype for rdf_fdw_table_clone function
This adds a prototype for the function rdf_fdw_table_clone, which
enables users to harvest all data (or a subset) from rdf_fdw foreign
tables. This function offers the possibility to extract data in small
chuncks to avoid large result sets from the target triplestore.

This patch also slightly changes the regression tests to a change in
the level of a FOREIGN TABLE logging option NOTICE > INFO.