Bugfix in the SQL SELECT parser
This fixes a bug that led some SPARQL queries to be created without
any variable in the SELECT clause. We now use * in case the planner
cannot identify which nodes should be retrieved, which can be a bit
inefficent if we're dealing with many columns, but it is better than
an error message.