Bug fix for xmldoc memory leak
The xml doc that contains the resulst set from the triple store
wasn't beeing freed after the query is complete. This could lead
to a memory leak that could cause a system crash once all available
was consumed - which is an issue for rather modest server settings
that execute mutliple queries in the same session.