Minor fixes to comparisons with rdfnode vs numeric types
This fixes a bug that led the comparisons of decimal-based XSD
rdfnodes with integer postgres types to fail, as the logic was
trying parse the lexical value of the rdfnode based on the type
on the other side of the expression.

This also removes the STRICT flag from the SPARQL functions that
return boolean - it now returns false for NULL parameters.

The SPARQL function CONCAT now accpets any number of rdfnodes
(variadic)

Regression tests were also updated to reflect minor formating
changes to the error and log messages.